+++ mne/decoding/time_gen.py
@@ -334,6 +334,7 @@ class _GeneralizationAcrossTime(object):
             number of testing times per training time need not be regular;
             else, np.shape(scores) = (n_train_time, n_test_time).
         """
+        import sklearn.metrics
         from sklearn.base import is_classifier
         from sklearn.metrics import accuracy_score, mean_squared_error
         if check_version('sklearn', '0.17'):
@@ -359,6 +360,14 @@ class _GeneralizationAcrossTime(object):
                     self.scorer_ = accuracy_score
                 elif is_regressor(self.clf):
                     self.scorer_ = mean_squared_error
+
+        elif isinstance(self.scorer_, str):
+            if hasattr(sklearn.metrics, '%s_score' % self.scorer_):
+                self.scorer_ = getattr(sklearn.metrics, '%s_score' %
+                                       self.scorer_)
+            else:
+                raise KeyError("`{0} scorer` Doesn't appear to be valid a "
+                               "scikit-learn scorer.".format(self.scorer_))
         if not self.scorer_:
             raise ValueError('Could not find a scoring metric for `clf=%s` '
                              ' and `predict_method=%s`. Manually define scorer'
@@ -897,8 +906,11 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
                 these predictions into a single estimate per sample.
 
         Default: 'cross-validation'
-    scorer : object | None
-        scikit-learn Scorer instance. If None, set to accuracy_score.
+    scorer : object | None | str
+        scikit-learn Scorer instance or str type indicating the name of
+        the scorer such as `accuracy`, `roc_auc`. If None, set to 
+        accuracy_score.
+
     n_jobs : int
         Number of jobs to run in parallel. Defaults to 1.
 
@@ -1195,8 +1207,10 @@ class TimeDecoding(_GeneralizationAcrossTime):
                 these predictions into a single estimate per sample.
 
         Default: 'cross-validation'
-    scorer : object | None
-        scikit-learn Scorer instance. If None, set to accuracy_score.
+    scorer : object | None | str
+        scikit-learn Scorer instance or str type indicating the
+        name of the scorer such as `accuracy`, `roc_auc`. If None,
+        set to accuracy_score.
     n_jobs : int
         Number of jobs to run in parallel. Defaults to 1.